refactor: 이미지 컨트롤러 통합
